refactor(api): move lcms to src/shared
