+++ api/src/shared/infrastructure/datasources/learning-content/datasource.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 
 import { LearningContentCache } from '../../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { lcms } from '../../../../../lib/infrastructure/lcms.js';
+import { lcms } from '../../lcms.js';
 import { LearningContentResourceNotFound } from './LearningContentResourceNotFound.js';
 
 const _DatasourcePrototype = {
