+++ api/tests/devcomp/unit/infrastructure/datasources/learning-content/tutorial-datasource_test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
 import { learningContentCache } from '../../../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { lcms } from '../../../../../../lib/infrastructure/lcms.js';
 import { tutorialDatasource } from '../../../../../../src/devcomp/infrastructure/datasources/learning-content/tutorial-datasource.js';
+import { lcms } from '../../../../../../src/shared/infrastructure/lcms.js';
 import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Datasource | Learning Content | TutorialDatasource', function () {
