+++ api/tests/shared/unit/infrastructure/datasources/learning-content/challenge-datasource_test.js
@@ -1,9 +1,9 @@
 import _ from 'lodash';
 
 import { learningContentCache } from '../../../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { lcms } from '../../../../../../lib/infrastructure/lcms.js';
 import { challengeDatasource } from '../../../../../../src/shared/infrastructure/datasources/learning-content/challenge-datasource.js';
 import { LearningContentResourceNotFound } from '../../../../../../src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
+import { lcms } from '../../../../../../src/shared/infrastructure/lcms.js';
 import { catchErr, expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Datasource | Learning Content | ChallengeDatasource', function () {
