+++ api/tests/shared/unit/infrastructure/datasources/learning-content/competence-datasource_test.js
@@ -1,6 +1,6 @@
 import { learningContentCache } from '../../../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { lcms } from '../../../../../../lib/infrastructure/lcms.js';
 import { competenceDatasource } from '../../../../../../src/shared/infrastructure/datasources/learning-content/competence-datasource.js';
+import { lcms } from '../../../../../../src/shared/infrastructure/lcms.js';
 import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Datasource | Learning Content | CompetenceDatasource', function () {
