+++ api/tests/shared/unit/infrastructure/datasources/learning-content/datasource_test.js
@@ -1,7 +1,7 @@
 import { learningContentCache } from '../../../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { lcms } from '../../../../../../lib/infrastructure/lcms.js';
 import * as dataSource from '../../../../../../src/shared/infrastructure/datasources/learning-content/datasource.js';
 import { LearningContentResourceNotFound } from '../../../../../../src/shared/infrastructure/datasources/learning-content/LearningContentResourceNotFound.js';
+import { lcms } from '../../../../../../src/shared/infrastructure/lcms.js';
 import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Datasource | Learning Content | datasource', function () {
