+++ api/tests/shared/unit/infrastructure/datasources/learning-content/tube-datasource_test.js
@@ -1,8 +1,8 @@
 import _ from 'lodash';
 
 import { learningContentCache } from '../../../../../../lib/infrastructure/caches/learning-content-cache.js';
-import { lcms } from '../../../../../../lib/infrastructure/lcms.js';
 import { tubeDatasource } from '../../../../../../src/shared/infrastructure/datasources/learning-content/tube-datasource.js';
+import { lcms } from '../../../../../../src/shared/infrastructure/lcms.js';
 import { expect, sinon } from '../../../../../test-helper.js';
 
 describe('Unit | Infrastructure | Datasource | Learning Content | TubeDatasource', function () {
